Changes for compilation.
